Consolidate no validation components
