+++ ui/packages/ui/src/app/components/index.ts
@@ -1,7 +1,10 @@
 export * from './ApplicationErrorPage';
 export * from './BasicSelectInput';
+export * from './FilterExcludeFieldComponent';
+export * from './FilterInputFieldComponent';
 export * from './FilterTreeComponent';
 export * from './KafkaConnectCluster';
+export * from './NoPreviewFilterField';
 export * from './PageLoader';
 export * from './PageNotFound';
 export * from './formHelpers';
