+++ ui/packages/ui/src/app/pages/createConnector/federatedModule/DataOptions.tsx
@@ -3,23 +3,25 @@ import {
   ExpandableSection,
   Form,
   Grid,
-  GridItem
+  GridItem,
+  Title
 } from "@patternfly/react-core";
 import { Formik } from "formik";
 import _ from "lodash";
 import React from "react";
-import { FormComponent } from "src/app/components/formHelpers";
-import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
-import { ConnectorNameTypeHeader } from "../../connectorSteps";
-import "./RuntimeOptions.css";
+import { FormComponent } from "../../../components/formHelpers";
+import { formatPropertyDefinitions, PropertyCategory } from "../../../shared";
+import { ConnectorNameTypeHeader } from "../connectorSteps";
+import "./DataOption.css";
 
-export interface IRuntimeOptionsProps {
+export interface IDataOptionsProps {
   connectorName: string;
   connectorType: string;
   configuration: Map<string, unknown>;
   propertyDefinitions: ConnectorProperty[];
-  i18nEngineProperties: string;
-  i18nHeartbeatProperties: string;
+  i18nAdvancedMappingPropertiesText: string;
+  i18nMappingPropertiesText: string;
+  i18nSnapshotPropertiesText: string;
   onChange: (configuration: Map<string, unknown>, isValid: boolean) => void;
 }
 
@@ -42,29 +44,38 @@ const checkIfRequired = (
   return matchProp ? matchProp.isMandatory : false;
 };
 
-export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
+export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
   const [initialValues, setInitialValues] = React.useState(
     getInitialObject(props.propertyDefinitions)
   );
-  const [engineExpanded, setEngineExpanded] = React.useState<boolean>(true);
-  const [heartbeatExpanded, setHeartbeatExpanded] = React.useState<boolean>(
-    true
-  );
+  const [mappingExpanded, setMappingExpanded] = React.useState<boolean>(false);
+  const [snapshotExpanded, setSnapshotExpanded] = React.useState<boolean>(true);
 
   const propertyDefinitionsCopy = _.cloneDeep(props.propertyDefinitions);
 
-  const enginePropertyDefinitions = formatPropertyDefinitions(
+  const mappingGeneralPropertyDefinitions = formatPropertyDefinitions(
     propertyDefinitionsCopy.filter(
-      (defn: ConnectorProperty) =>
-        defn.category === PropertyCategory.RUNTIME_OPTIONS_ENGINE
+      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_GENERAL
     )
   );
-  const heartbeatPropertyDefinitions = formatPropertyDefinitions(
+  const mappingAdvancedPropertyDefinitions = formatPropertyDefinitions(
     propertyDefinitionsCopy.filter(
-      (defn: ConnectorProperty) =>
-        defn.category === PropertyCategory.RUNTIME_OPTIONS_HEARTBEAT
+      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_ADVANCED
     )
   );
+  const snapshotPropertyDefinitions = formatPropertyDefinitions(
+    propertyDefinitionsCopy.filter(
+      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_SNAPSHOT
+    )
+  );
+
+  const onToggleMapping = (isExpanded: boolean) => {
+    setMappingExpanded(isExpanded);
+  };
+
+  const onToggleSnapshot = (isExpanded: boolean) => {
+    setSnapshotExpanded(isExpanded);
+  };
 
   const validateForm = (values: any) => {
     const formEntries = Object.entries(values).reduce((a,[k,v])=>(initialValues[k]===v||(a[k]=v),a),{});
@@ -99,14 +110,6 @@ export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
     return isValid;
   };
 
-  const onToggleEngine = (isExpanded: boolean) => {
-    setEngineExpanded(isExpanded);
-  };
-
-  const onToggleHeartbeat = (isExpanded: boolean) => {
-    setHeartbeatExpanded(isExpanded);
-  };
-
   const handlePropertyChange = (propName: string, propValue: any) => {
     // TODO: handling for property change if needed.
   };
@@ -133,10 +136,7 @@ export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
   }, []);
 
   return (
-    <div
-      style={{ padding: "20px" }}
-      className={"runtime-options-component-page"}
-    >
+    <div style={{ padding: "20px" }} className={"data-options-component-page"}>
       <ConnectorNameTypeHeader
         connectorName={props.connectorName}
         connectorType={props.connectorType}
@@ -158,16 +158,18 @@ export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
                 <GridItem lg={9} sm={12}>
                   <ExpandableSection
                     toggleText={
-                      engineExpanded ? props.i18nEngineProperties : props.i18nEngineProperties
+                      snapshotExpanded
+                        ? props.i18nSnapshotPropertiesText
+                        : props.i18nSnapshotPropertiesText
                     }
-                    onToggle={onToggleEngine}
-                    isExpanded={engineExpanded}
+                    onToggle={onToggleSnapshot}
+                    isExpanded={snapshotExpanded}
                   >
                     <Grid
                       hasGutter={true}
-                      className={"runtime-options-component-expansion-content"}
+                      className={"data-options-component-expansion-content"}
                     >
-                      {enginePropertyDefinitions
+                      {snapshotPropertyDefinitions
                         .map((propertyDefinition: ConnectorProperty, index) => {
                           return (
                             <GridItem
@@ -189,18 +191,47 @@ export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
                   </ExpandableSection>
                   <ExpandableSection
                     toggleText={
-                      heartbeatExpanded
-                        ? props.i18nHeartbeatProperties
-                        : props.i18nHeartbeatProperties
+                      mappingExpanded
+                        ? props.i18nMappingPropertiesText
+                        : props.i18nMappingPropertiesText
                     }
-                    onToggle={onToggleHeartbeat}
-                    isExpanded={heartbeatExpanded}
+                    onToggle={onToggleMapping}
+                    isExpanded={mappingExpanded}
                   >
                     <Grid
                       hasGutter={true}
-                      className={"runtime-options-component-expansion-content"}
+                      className={"data-options-component-expansion-content"}
                     >
-                      {heartbeatPropertyDefinitions
+                      {mappingGeneralPropertyDefinitions
+                        .map((propertyDefinition: ConnectorProperty, index) => {
+                          return (
+                            <GridItem
+                              key={index}
+                              lg={propertyDefinition.gridWidthLg}
+                              sm={propertyDefinition.gridWidthSm}
+                            >
+                              <FormComponent
+                                propertyDefinition={propertyDefinition}
+                                propertyChange={handlePropertyChange}
+                                setFieldValue={setFieldValue}
+                                invalidMsg={[]}
+                                validated={"default"}
+                              />
+                            </GridItem>
+                          );
+                        })}
+                    </Grid>
+                    <Title
+                      headingLevel="h3"
+                      className={"data-options-component-grouping"}
+                    >
+                      {props.i18nAdvancedMappingPropertiesText}
+                    </Title>
+                    <Grid
+                      hasGutter={true}
+                      className={"data-options-component-expansion-content"}
+                    >
+                      {mappingAdvancedPropertyDefinitions
                         .map((propertyDefinition: ConnectorProperty, index) => {
                           return (
                             <GridItem
