+++ ui/packages/ui/src/app/pages/createConnector/federatedModule/RuntimeOptions.tsx
@@ -3,25 +3,23 @@ import {
   ExpandableSection,
   Form,
   Grid,
-  GridItem,
-  Title
+  GridItem
 } from "@patternfly/react-core";
 import { Formik } from "formik";
 import _ from "lodash";
 import React from "react";
-import { FormComponent } from "src/app/components/formHelpers";
-import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
-import { ConnectorNameTypeHeader } from "../../connectorSteps";
-import "./DataOption.css";
+import { FormComponent } from "../../../components/formHelpers";
+import { formatPropertyDefinitions, PropertyCategory } from "../../../shared";
+import { ConnectorNameTypeHeader } from "../connectorSteps";
+import "./RuntimeOptions.css";
 
-export interface IDataOptionsProps {
+export interface IRuntimeOptionsProps {
   connectorName: string;
   connectorType: string;
   configuration: Map<string, unknown>;
   propertyDefinitions: ConnectorProperty[];
-  i18nAdvancedMappingPropertiesText: string;
-  i18nMappingPropertiesText: string;
-  i18nSnapshotPropertiesText: string;
+  i18nEngineProperties: string;
+  i18nHeartbeatProperties: string;
   onChange: (configuration: Map<string, unknown>, isValid: boolean) => void;
 }
 
@@ -44,38 +42,29 @@ const checkIfRequired = (
   return matchProp ? matchProp.isMandatory : false;
 };
 
-export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
+export const RuntimeOptions: React.FC<IRuntimeOptionsProps> = (props) => {
   const [initialValues, setInitialValues] = React.useState(
     getInitialObject(props.propertyDefinitions)
   );
-  const [mappingExpanded, setMappingExpanded] = React.useState<boolean>(false);
-  const [snapshotExpanded, setSnapshotExpanded] = React.useState<boolean>(true);
+  const [engineExpanded, setEngineExpanded] = React.useState<boolean>(true);
+  const [heartbeatExpanded, setHeartbeatExpanded] = React.useState<boolean>(
+    true
+  );
 
   const propertyDefinitionsCopy = _.cloneDeep(props.propertyDefinitions);
 
-  const mappingGeneralPropertyDefinitions = formatPropertyDefinitions(
+  const enginePropertyDefinitions = formatPropertyDefinitions(
     propertyDefinitionsCopy.filter(
-      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_GENERAL
+      (defn: ConnectorProperty) =>
+        defn.category === PropertyCategory.RUNTIME_OPTIONS_ENGINE
     )
   );
-  const mappingAdvancedPropertyDefinitions = formatPropertyDefinitions(
+  const heartbeatPropertyDefinitions = formatPropertyDefinitions(
     propertyDefinitionsCopy.filter(
-      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_ADVANCED
+      (defn: ConnectorProperty) =>
+        defn.category === PropertyCategory.RUNTIME_OPTIONS_HEARTBEAT
     )
   );
-  const snapshotPropertyDefinitions = formatPropertyDefinitions(
-    propertyDefinitionsCopy.filter(
-      (defn: any) => defn.category === PropertyCategory.DATA_OPTIONS_SNAPSHOT
-    )
-  );
-
-  const onToggleMapping = (isExpanded: boolean) => {
-    setMappingExpanded(isExpanded);
-  };
-
-  const onToggleSnapshot = (isExpanded: boolean) => {
-    setSnapshotExpanded(isExpanded);
-  };
 
   const validateForm = (values: any) => {
     const formEntries = Object.entries(values).reduce((a,[k,v])=>(initialValues[k]===v||(a[k]=v),a),{});
@@ -110,6 +99,14 @@ export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
     return isValid;
   };
 
+  const onToggleEngine = (isExpanded: boolean) => {
+    setEngineExpanded(isExpanded);
+  };
+
+  const onToggleHeartbeat = (isExpanded: boolean) => {
+    setHeartbeatExpanded(isExpanded);
+  };
+
   const handlePropertyChange = (propName: string, propValue: any) => {
     // TODO: handling for property change if needed.
   };
@@ -136,7 +133,10 @@ export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
   }, []);
 
   return (
-    <div style={{ padding: "20px" }} className={"data-options-component-page"}>
+    <div
+      style={{ padding: "20px" }}
+      className={"runtime-options-component-page"}
+    >
       <ConnectorNameTypeHeader
         connectorName={props.connectorName}
         connectorType={props.connectorType}
@@ -158,18 +158,16 @@ export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
                 <GridItem lg={9} sm={12}>
                   <ExpandableSection
                     toggleText={
-                      snapshotExpanded
-                        ? props.i18nSnapshotPropertiesText
-                        : props.i18nSnapshotPropertiesText
+                      engineExpanded ? props.i18nEngineProperties : props.i18nEngineProperties
                     }
-                    onToggle={onToggleSnapshot}
-                    isExpanded={snapshotExpanded}
+                    onToggle={onToggleEngine}
+                    isExpanded={engineExpanded}
                   >
                     <Grid
                       hasGutter={true}
-                      className={"data-options-component-expansion-content"}
+                      className={"runtime-options-component-expansion-content"}
                     >
-                      {snapshotPropertyDefinitions
+                      {enginePropertyDefinitions
                         .map((propertyDefinition: ConnectorProperty, index) => {
                           return (
                             <GridItem
@@ -191,47 +189,18 @@ export const DataOptions: React.FC<IDataOptionsProps> = (props) => {
                   </ExpandableSection>
                   <ExpandableSection
                     toggleText={
-                      mappingExpanded
-                        ? props.i18nMappingPropertiesText
-                        : props.i18nMappingPropertiesText
+                      heartbeatExpanded
+                        ? props.i18nHeartbeatProperties
+                        : props.i18nHeartbeatProperties
                     }
-                    onToggle={onToggleMapping}
-                    isExpanded={mappingExpanded}
+                    onToggle={onToggleHeartbeat}
+                    isExpanded={heartbeatExpanded}
                   >
                     <Grid
                       hasGutter={true}
-                      className={"data-options-component-expansion-content"}
+                      className={"runtime-options-component-expansion-content"}
                     >
-                      {mappingGeneralPropertyDefinitions
-                        .map((propertyDefinition: ConnectorProperty, index) => {
-                          return (
-                            <GridItem
-                              key={index}
-                              lg={propertyDefinition.gridWidthLg}
-                              sm={propertyDefinition.gridWidthSm}
-                            >
-                              <FormComponent
-                                propertyDefinition={propertyDefinition}
-                                propertyChange={handlePropertyChange}
-                                setFieldValue={setFieldValue}
-                                invalidMsg={[]}
-                                validated={"default"}
-                              />
-                            </GridItem>
-                          );
-                        })}
-                    </Grid>
-                    <Title
-                      headingLevel="h3"
-                      className={"data-options-component-grouping"}
-                    >
-                      {props.i18nAdvancedMappingPropertiesText}
-                    </Title>
-                    <Grid
-                      hasGutter={true}
-                      className={"data-options-component-expansion-content"}
-                    >
-                      {mappingAdvancedPropertyDefinitions
+                      {heartbeatPropertyDefinitions
                         .map((propertyDefinition: ConnectorProperty, index) => {
                           return (
                             <GridItem
