+++ ui/packages/ui/src/app/pages/createConnector/federatedModule/config.ts
