--- ui/packages/ui/src/app/pages/createConnector/noValidation/CreateConnectorNoValidation.css
@@ -1,69 +1,0 @@
-.pf-c-page__main {
-  overflow: unset;
-}
-.create-connector {
-  &-page_breadcrumb {
-    padding: 0px !important;
-    border-bottom: 1px solid var(--pf-global--BorderColor--100);
-    .pf-l-level {
-      .pf-c-content {
-        padding: 15px 0px;
-      }
-    }
-  }
-
-  &-page_wizard {
-    height: auto;
-    min-height: calc(100vh - 200px);
-
-    .pf-c-wizard {
-      &__main{
-        overflow: unset;
-      }
-      &__nav-link{
-        &.pf-m-current{
-          color: var(--pf-c-wizard__nav-link--m-current--Color);
-        }
-      }
-      &__nav-list {
-        margin: 24px 5px;
-        padding: 0 10px 0 30px;
-      }
-      &__nav-item {
-        margin-bottom: 30px;
-        ol{
-          padding: 0 10px;
-          li{
-            margin-bottom: 20px;
-          }
-        }
-        .pf-m-disabled {
-          &:before {
-            background-color: var(--pf-global--disabled-color--300);
-          }
-        }
-        .pf-m-required{
-          color: var(--pf-global--palette--red-100);
-        }
-      }
-    }
-  }
-}
-
-.skipToNextInfoBox{
-  width: 100%;
-  margin-bottom: 40px;
-  .pf-c-button.pf-m-link{
-    padding-left: 5px;
-    padding-right: 5px;
-  }
-  .titleSuccess{
-    color: var(--pf-global--palette--green-500);
-  }
-  .titleInfo{
-    color: var(--pf-global--info-color--100);
-  }
-}
-.pf-c-notification-drawer{
-  box-shadow: var(--pf-global--BoxShadow--sm);
-}
