--- ui/packages/ui/src/app/pages/createConnector/noValidation/connectorSteps/DataOptionsNoValidationStep.css
@@ -1,42 +1,0 @@
-.c-data-options {
-  &__spacer {
-    margin: 20px 0;
-  }
-}
-
-.data-options-component {
-  &-grouping {
-    margin-top: 30px;
-    margin-bottom: 10px;
-    margin-left: 30px;
-  }
-  &-mapping-title {
-    margin-top: 30px;
-  }
-  &-expansion-content {
-    @media screen and (min-width: 768px) {
-      padding-left: 30px;
-    }
-  }
-}
-.data-options-component-page {
-  .pf-c-expandable-section {
-    padding-bottom: 30px;
-    &__toggle {
-      justify-content: flex-start;
-      border-bottom: var(--pf-global--BorderWidth--sm) solid #e6e6e6 !important;
-      width: 100%;
-
-      &-text {
-        font-size: var(--pf-global--FontSize--xl);
-        font-weight: var(--pf-global--FontWeight--bold);
-        color: var(--pf-global--Color--100);
-      }
-    }
-  }
-  .pf-c-switch, .pf-c-switch__label {
-    @media screen and (max-width: 768px) {
-      display: inline;
-    }
-  }
-}
