--- ui/packages/ui/src/app/pages/createConnector/noValidation/connectorSteps/PropertiesStep.css
@@ -1,39 +1,0 @@
-.connector-name-form {
-  padding: 8px 0px;
-}
-.connector-type-label {
-  font-weight: var(--pf-c-form__label-text--FontWeight);
-  font-size: var(--pf-c-form__label--FontSize);
-  margin-right: 10px;
-}
-.properties-step {
-  &-advanced-title {
-    margin-top: 30px;
-  }
-  &-grouping {
-    margin-top: 30px;
-    margin-bottom: 10px;
-    margin-left: 30px;
-  }
-  &-expansion-content {
-    @media screen and (min-width: 768px) {
-      padding-left: 30px;
-    }
-  }
-}
-.properties-step-page {
-  .pf-c-expandable-section {
-    padding-bottom: 30px;
-    &__toggle {
-      justify-content: flex-start;
-      border-bottom: var(--pf-global--BorderWidth--sm) solid #e6e6e6 !important;
-      width: 100%;
-
-      &-text {
-        font-size: var(--pf-global--FontSize--xl);
-        font-weight: var(--pf-global--FontWeight--bold);
-        color: var(--pf-global--Color--100);
-      }
-    }
-  }
-}
