--- ui/packages/ui/src/app/pages/createConnector/noValidation/connectorSteps/RuntimeOptionsNoValidationStep.css
@@ -1,32 +1,0 @@
-.c-runtime-options {
-  &__spacer {
-    margin: 20px 0;
-  }
-}
-
-.runtime-options-component {
-  &-heartbeat-title {
-    margin-top: 30px;
-  }
-  &-expansion-content {
-    @media screen and (min-width: 768px) {
-      padding-left: 30px;
-    }
-  }
-}
-.runtime-options-component-page {
-  .pf-c-expandable-section {
-    padding-bottom: 30px;
-    &__toggle {
-      justify-content: flex-start;
-      border-bottom: var(--pf-global--BorderWidth--sm) solid #e6e6e6 !important;
-      width: 100%;
-
-      &-text {
-        font-size: var(--pf-global--FontSize--xl);
-        font-weight: var(--pf-global--FontWeight--bold);
-        color: var(--pf-global--Color--100);
-      }
-    }
-  }
-}
