--- ui/packages/ui/src/app/pages/createConnector/noValidation/connectorSteps/index.ts
@@ -1,4 +1,0 @@
-export * from './FilterConfigNoValidationStep';
-export * from './PropertiesStep';
-export * from './DataOptionsNoValidationStep';
-export * from './RuntimeOptionsNoValidationStep';
